backend/requirements/common.txt
requests==2.2.0
flask==1.0.2
flask_restful==0.3.7
flask-sqlalchemy==2.3.2
flask-cors==3.0.7
psycopg2==2.7.7